+++ eve/Week8/login/src/hooks/useSidebar.ts
@@ -1,13 +1,29 @@
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 
 export const useSidebar = (initialState = false) => {
   const [isOpen, setIsOpen] = useState(initialState);
 
   const open = () => setIsOpen(true);
-
   const close = () => setIsOpen(false);
-
   const toggle = () => setIsOpen(!isOpen);
+
+  useEffect(() => {
+    if (!isOpen) return; 
+
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (e.key === 'Escape') {
+        close();
+      }
+    };
+
+    document.body.style.overflow = 'hidden';
+    window.addEventListener('keydown', handleKeyDown);
+
+    return () => {
+      document.body.style.overflow = 'unset';
+      window.removeEventListener('keydown', handleKeyDown);
+    };
+  }, [isOpen]);
 
   return {
     isOpen,
